+++ pennylane/ops/qubit.py
@@ -15,12 +15,17 @@
 This module contains the available built-in discrete-variable
 quantum operations supported by PennyLane, as well as their conventions.
 """
-# pylint:disable=abstract-method,arguments-differ
+# pylint:disable=abstract-method,arguments-differ,protected-access
 import numpy as np
+from scipy.linalg import block_diag
 
 from pennylane.operation import Any, Observable, Operation
-from pennylane.templates.state_preparations import BasisStatePreparation, MottonenStatePreparation
+from pennylane.templates.state_preparations import (
+    BasisStatePreparation,
+    MottonenStatePreparation,
+)
 from pennylane.utils import OperationRecorder, pauli_eigs
+
 
 class Hadamard(Observable, Operation):
     r"""Hadamard(wires)
@@ -40,6 +45,26 @@ class Hadamard(Observable, Operation):
     num_wires = 1
     par_domain = None
     eigvals = pauli_eigs(1)
+
+    @staticmethod
+    def _matrix(*params):
+        return np.array([[1, 1], [1, -1]]) / np.sqrt(2)
+
+    def diagonalizing_gates(self):
+        r"""Rotates the specified wires such that they
+        are in the eigenbasis of the Hadamard operator.
+
+        For the Hadamard operator,
+
+        .. math:: H = U^\dagger Z U
+
+        where :math:`U = R_y(-\pi/4)`.
+
+        Returns:
+            list(qml.Operation): A list of gates that diagonalize Hadamard in the
+            computational basis.
+        """
+        return [RY(-np.pi/4, wires=self.wires)]
 
 
 class PauliX(Observable, Operation):
@@ -61,6 +86,24 @@ class PauliX(Observable, Operation):
     par_domain = None
     eigvals = pauli_eigs(1)
 
+    @staticmethod
+    def _matrix(*params):
+        return np.array([[0, 1], [1, 0]])
+
+    def diagonalizing_gates(self):
+        r"""Rotates the specified wires such that they
+        are in the eigenbasis of the Pauli-X operator.
+
+        For the Pauli-X operator,
+
+        .. math:: X = H^\dagger Z H.
+
+        Returns:
+            list(qml.Operation): A list of gates that diagonalize PauliY in the
+            computational basis.
+        """
+        return [Hadamard(wires=self.wires)]
+
 
 class PauliY(Observable, Operation):
     r"""PauliY(wires)
@@ -81,6 +124,26 @@ class PauliY(Observable, Operation):
     par_domain = None
     eigvals = pauli_eigs(1)
 
+    @staticmethod
+    def _matrix(*params):
+        return np.array([[0, -1j], [1j, 0]])
+
+    def diagonalizing_gates(self):
+        r"""Rotates the specified wires such that they
+        are in the eigenbasis of PauliY.
+
+        For the Pauli-Y observable,
+
+        .. math:: Y = U^\dagger Z U
+
+        where :math:`U=HSZ`.
+
+        Returns:
+            list(qml.Operation): A list of gates that diagonalize PauliY in the
+            computational basis.
+        """
+        return [Hadamard(wires=self.wires), S(wires=self.wires), PauliZ(wires=self.wires)]
+
 
 class PauliZ(Observable, Operation):
     r"""PauliZ(wires)
@@ -100,6 +163,13 @@ class PauliZ(Observable, Operation):
     num_wires = 1
     par_domain = None
     eigvals = pauli_eigs(1)
+
+    @staticmethod
+    def _matrix(*params):
+        return np.array([[1, 0], [0, -1]])
+
+    def diagonalizing_gates(self):
+        return []
 
 
 class S(Operation):
@@ -123,6 +193,10 @@ class S(Operation):
     num_wires = 1
     par_domain = None
 
+    @staticmethod
+    def _matrix(*params):
+        return np.array([[1, 0], [0, 1j]])
+
 
 class T(Operation):
     r"""T(wires)
@@ -144,6 +218,10 @@ class T(Operation):
     num_params = 0
     num_wires = 1
     par_domain = None
+
+    @staticmethod
+    def _matrix(*params):
+        return np.array([[1, 0], [0, np.exp(1j*np.pi/4)]])
 
 
 class CNOT(Operation):
@@ -171,6 +249,10 @@ class CNOT(Operation):
     num_wires = 2
     par_domain = None
 
+    @staticmethod
+    def _matrix(*params):
+        return np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]])
+
 
 class CZ(Operation):
     r"""CZ(wires)
@@ -197,6 +279,10 @@ class CZ(Operation):
     num_wires = 2
     par_domain = None
 
+    @staticmethod
+    def _matrix(*params):
+        return np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, -1]])
+
 
 class SWAP(Operation):
     r"""SWAP(wires)
@@ -220,6 +306,10 @@ class SWAP(Operation):
     num_params = 0
     num_wires = 2
     par_domain = None
+
+    @staticmethod
+    def _matrix(*params):
+        return np.array([[1, 0, 0, 0], [0, 0, 1, 0], [0, 1, 0, 0], [0, 0, 0, 1]])
 
 
 class CSWAP(Operation):
@@ -250,6 +340,17 @@ class CSWAP(Operation):
     num_params = 0
     num_wires = 3
     par_domain = None
+
+    @staticmethod
+    def _matrix(*params):
+        return np.array([[1, 0, 0, 0, 0, 0, 0, 0],
+                        [0, 1, 0, 0, 0, 0, 0, 0],
+                        [0, 0, 1, 0, 0, 0, 0, 0],
+                        [0, 0, 0, 1, 0, 0, 0, 0],
+                        [0, 0, 0, 0, 1, 0, 0, 0],
+                        [0, 0, 0, 0, 0, 0, 1, 0],
+                        [0, 0, 0, 0, 0, 1, 0, 0],
+                        [0, 0, 0, 0, 0, 0, 0, 1]])
 
 
 class Toffoli(Operation):
@@ -282,6 +383,12 @@ class Toffoli(Operation):
     num_wires = 3
     par_domain = None
 
+    @staticmethod
+    def _matrix(*params):
+        mat = np.diag([1 for i in range(8)])
+        mat[6:8, 6:8] = np.array([[0, 1], [1, 0]])
+        return mat
+
 
 class RX(Operation):
     r"""RX(phi, wires)
@@ -308,6 +415,11 @@ class RX(Operation):
     par_domain = "R"
     grad_method = "A"
     generator = [PauliX, -1 / 2]
+
+    @staticmethod
+    def _matrix(*params):
+        theta = params[0]
+        return np.cos(theta/2) * np.eye(2) + 1j * np.sin(-theta/2) * PauliX._matrix()
 
 
 class RY(Operation):
@@ -336,6 +448,11 @@ class RY(Operation):
     grad_method = "A"
     generator = [PauliY, -1 / 2]
 
+    @staticmethod
+    def _matrix(*params):
+        theta = params[0]
+        return np.cos(theta/2) * np.eye(2) + 1j * np.sin(-theta/2) * PauliY._matrix()
+
 
 class RZ(Operation):
     r"""RZ(phi, wires)
@@ -363,6 +480,11 @@ class RZ(Operation):
     grad_method = "A"
     generator = [PauliZ, -1 / 2]
 
+    @staticmethod
+    def _matrix(*params):
+        theta = params[0]
+        return np.cos(theta/2) * np.eye(2) + 1j * np.sin(-theta/2) * PauliZ._matrix()
+
 
 class PhaseShift(Operation):
     r"""PhaseShift(phi, wires)
@@ -389,6 +511,11 @@ class PhaseShift(Operation):
     par_domain = "R"
     grad_method = "A"
     generator = [np.array([[0, 0], [0, 1]]), 1]
+
+    @staticmethod
+    def _matrix(*params):
+        phi = params[0]
+        return np.array([[1, 0], [0, np.exp(1j*phi)]])
 
 
 class Rot(Operation):
@@ -427,11 +554,16 @@ class Rot(Operation):
     grad_method = "A"
 
     @staticmethod
+    def _matrix(*params):
+        a, b, c = params
+        return RZ._matrix(c) @ (RY._matrix(b) @ RZ._matrix(a))
+
+    @staticmethod
     def decomposition(phi, theta, omega, wires):
         decomp_ops = [
             RZ(phi, wires=wires),
             RY(theta, wires=wires),
-            RZ(omega, wires=wires)
+            RZ(omega, wires=wires),
         ]
         return decomp_ops
 
@@ -482,7 +614,14 @@ class CRX(Operation):
     num_wires = 2
     par_domain = "R"
     grad_method = "A"
-    generator = [np.array([[0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]]), -1 / 2]
+    generator = [
+        np.array([[0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]]),
+        -1 / 2,
+    ]
+
+    @staticmethod
+    def _matrix(*params):
+        return block_diag(np.eye(2), RX._matrix(*params))
 
     @staticmethod
     def decomposition(theta, wires):
@@ -541,7 +680,14 @@ class CRY(Operation):
     num_wires = 2
     par_domain = "R"
     grad_method = "A"
-    generator = [np.array([[0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 0, -1j], [0, 0, 1j, 0]]), -1 / 2]
+    generator = [
+        np.array([[0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 0, -1j], [0, 0, 1j, 0]]),
+        -1 / 2,
+    ]
+
+    @staticmethod
+    def _matrix(*params):
+        return block_diag(np.eye(2), RY._matrix(*params))
 
     @staticmethod
     def decomposition(theta, wires):
@@ -549,7 +695,7 @@ class CRY(Operation):
             U3(theta / 2, 0, 0, wires=wires[1]),
             CNOT(wires=wires),
             U3(-theta / 2, 0, 0, wires=wires[1]),
-            CNOT(wires=wires)
+            CNOT(wires=wires),
         ]
         return decomp_ops
 
@@ -598,7 +744,14 @@ class CRZ(Operation):
     num_wires = 2
     par_domain = "R"
     grad_method = "A"
-    generator = [np.array([[0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 1, 0], [0, 0, 0, -1]]), -1 / 2]
+    generator = [
+        np.array([[0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 1, 0], [0, 0, 0, -1]]),
+        -1 / 2,
+    ]
+
+    @staticmethod
+    def _matrix(*params):
+        return block_diag(np.eye(2), RZ._matrix(*params))
 
     @staticmethod
     def decomposition(lam, wires):
@@ -606,7 +759,7 @@ class CRZ(Operation):
             PhaseShift(lam / 2, wires=wires[1]),
             CNOT(wires=wires),
             PhaseShift(-lam / 2, wires=wires[1]),
-            CNOT(wires=wires)
+            CNOT(wires=wires),
         ]
         return decomp_ops
 
@@ -643,6 +796,11 @@ class CRot(Operation):
     par_domain = "R"
     grad_method = "A"
 
+    @staticmethod
+    def _matrix(*params):
+        a, b, c = params
+        return CRZ._matrix(c) @ (CRY._matrix(b) @ CRZ._matrix(a))
+
 
 class U1(Operation):
     r"""U1(phi)
@@ -673,6 +831,10 @@ class U1(Operation):
     par_domain = "R"
     grad_method = "A"
     generator = [np.array([[0, 0], [0, 1]]), 1]
+
+    @staticmethod
+    def _matrix(*params):
+        return PhaseShift._matrix(*params)
 
     @staticmethod
     def decomposition(phi, wires):
@@ -717,6 +879,11 @@ class U2(Operation):
     num_wires = 1
     par_domain = "R"
     grad_method = "A"
+
+    @staticmethod
+    def _matrix(*params):
+        phi, lam = params
+        return PhaseShift._matrix(phi+lam) @ Rot._matrix(lam, np.pi/2, -lam)
 
     @staticmethod
     def decomposition(phi, lam, wires):
@@ -768,6 +935,12 @@ class U3(Operation):
     par_domain = "R"
     grad_method = "A"
 
+
+    @staticmethod
+    def _matrix(*params):
+        theta, phi, lam = params
+        return PhaseShift._matrix(phi+lam) @ Rot._matrix(lam, theta, -lam)
+
     @staticmethod
     def decomposition(theta, phi, lam, wires):
         decomp_ops = [
@@ -801,6 +974,18 @@ class QubitUnitary(Operation):
     num_wires = Any
     par_domain = "A"
     grad_method = None
+
+    @staticmethod
+    def _matrix(*params):
+        U = np.asarray(params[0])
+
+        if U.shape[0] != U.shape[1]:
+            raise ValueError("Operator must be a square matrix.")
+
+        if not np.allclose(U @ U.conj().T, np.identity(U.shape[0])):
+            raise ValueError("Operator must be unitary.")
+
+        return U
 
 
 # =============================================================================
@@ -912,8 +1097,20 @@ class Hermitian(Observable):
     grad_method = "F"
     _eigs = {}
 
-    @classmethod
-    def eigvals(cls, Hmat):
+    @staticmethod
+    def _matrix(*params):
+        A = np.asarray(params[0])
+
+        if A.shape[0] != A.shape[1]:
+            raise ValueError("Observable must be a square matrix.")
+
+        if not np.allclose(A, A.conj().T):
+            raise ValueError("Observable must be Hermitian.")
+
+        return A
+
+    @property
+    def eigvals(self):
         """Return the eigenvalues of the specified Hermitian observable.
 
         This method uses pre-stored eigenvalues for standard observables where
@@ -922,15 +1119,14 @@ class Hermitian(Observable):
         Returns:
             array: array containing the eigenvalues of the Hermitian observable
         """
-        Hmat = reshape_square_operator(Hmat)
-        Hkey = tuple(Hmat.flatten().tolist())
-        if Hkey not in cls._eigs:
+        Hmat = self.matrix
+        Hkey = tuple(self.parameters[0].flatten().tolist())
+        if Hkey not in Hermitian._eigs:
             w, U = np.linalg.eigh(Hmat)
-            cls._eigs[Hkey] = {"eigval": w, "eigvec": U}
-        return cls._eigs[Hkey]["eigval"]
+            Hermitian._eigs[Hkey] = {"eigval": w, "eigvec": U}
+        return Hermitian._eigs[Hkey]["eigval"]
 
-    @classmethod
-    def diagonalizing_gates(cls, Hmat, wires):
+    def diagonalizing_gates(self):
         """Return the gate set that diagonalizes a circuit according to the
         specified Hermitian observable.
 
@@ -938,18 +1134,15 @@ class Hermitian(Observable):
         possible and stores the corresponding eigenvectors from the eigendecomposition.
 
         Returns:
-            array: array containing the eigenvalues of the tensor product observable
+            list: list containing the gates diagonalizing the Hermitian observable
         """
-        Hmat = reshape_square_operator(Hmat)
+        Hmat = self.matrix
         Hkey = tuple(Hmat.flatten().tolist())
-        if Hkey not in cls._eigs:
+        if Hkey not in Hermitian._eigs:
             w, U = np.linalg.eigh(Hmat)
-            cls._eigs[Hkey] = {"eigval": w, "eigvec": U}
-        U = cls._eigs[Hkey]["eigvec"]
-        U_dim = int(np.sqrt(np.prod(U.shape)))
-
+            Hermitian._eigs[Hkey] = {"eigval": w, "eigvec": U}
         return [
-            QubitUnitary(cls._eigs[Hkey]["eigvec"].reshape(U_dim, U_dim).conj().T, wires=wires),
+            QubitUnitary(Hermitian._eigs[Hkey]["eigvec"].conj().T, wires=self.wires),
         ]